+++ utils.c
@@ -15,11 +15,11 @@
 
 /* returns bytes or unicode internal buffer */
 static PyObject*
-pymod_get_string(PyObject* obj, DAWG_LETTER_TYPE** word, ssize_t* wordlen) {
+pymod_get_string(PyObject* obj, DAWG_LETTER_TYPE** word, size_t* wordlen) {
 #ifdef DAWG_UNICODE
 	if (PyUnicode_Check(obj)) {
 		*word = (DAWG_LETTER_TYPE*)PyUnicode_AS_UNICODE(obj);
-		*wordlen = PyUnicode_GET_SIZE(obj);
+		*wordlen = (size_t)PyUnicode_GET_SIZE(obj);
 		Py_INCREF(obj);
 		return obj;
 	}
@@ -30,7 +30,7 @@ pymod_get_string(PyObject* obj, DAWG_LETTER_TYPE** word, ssize_t* wordlen) {
 #else
 	if (PyBytes_Check(obj)) {
 		*word = (uint8_t*)PyBytes_AS_STRING(obj);
-		*wordlen = PyBytes_GET_SIZE(obj);
+		*wordlen = (size_t)PyBytes_GET_SIZE(obj);
 		Py_INCREF(obj);
 		return obj;
 	}
